ai-accountant-kz/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# AI
openai==1.10.0
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2
email-validator==2.1.0

# Tasks
celery==5.3.4
redis==5.0.1

# Telegram Bot
aiogram==3.4.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# PDF Generation
reportlab==4.0.7

# XML
lxml==4.9.3

# OCR (for receipts)
pytesseract==0.3.10
pillow==10.1.0
